+++ CHANGELOG.md
@@ -14,3 +14,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated `getAuthorization` to use the correct API URL.
 - Rename `getConnection(name: string)` -> `getAuthorization(developerName: string, attachmentNameOrColorUrl = "HEROKU_APPLINK")`, accepting a new attachmentNameOrColorOrUrl to use a specific Applink addon's config.
 - Remove node-fetch in favor of native fetch, add `HTTPResponseError`
+
+## Unreleased
+- Add `X-App-UUID` header, `heroku-applink-node-sdk` UA.